fixes an issue where negative indices could get cached thus causing them to no longer refer to their calculated index.
also fixes the performance of the repr methods in both account managers, as it was very slow with some experimental account plugins installed.

basically just a bunch of fixes to test-accounts that i found when working on the new `ape-titanoboa` plugin.

fixes: #2436 
Fixes: APE-1865

```
BEFORE:

In [1]: %time accounts.test_accounts
Wall time: 17.1 ms
Wall time: 10 μs

In [1]: %time accounts.test_accounts[5]
Wall time: 329 ms
Wall time: 30 μs

In [1]: %time accounts.test_accounts[0].address
Wall time: 344 ms
Wall time: 250 μs


AFTER:

In [1]: %time accounts.test_accounts
Wall time: 10.4 ms
Wall time: 7.87 μs

In [1]: %time accounts.test_accounts[5]
Wall time: 335 ms
Wall time: 34.1 μs

In [1]: %time accounts.test_accounts[0].address
Wall time: 325 ms
Wall time: 38.9 μs
```